requirements.txt
streamlit==1.19.0
fastapi
uvicorn
python-pptx
python-bidi
arabic-reshaper
pillow
python-dotenv
openai
fpdf
transformers 
peft 
datasets 
accelerate 
pandas 
openpyxl
sentencepiece
sacremoses
altair==4
